debug decrease du panier - Test 1

+++ ap_hero/src/Service/Cart/CartService.php
@@ -86,6 +86,7 @@ class CartService
             $cartItem = new CartItem();
             $cartItem->setProduct($this->variantRepository->find($id));
             $cartItem->setQuantity($quantity);
+            $cartItem->setIsPaid(false);
             $cartEntity->addCartItems($cartItem);
             $this->entityManager->persist($cartItem);
         }
@@ -106,6 +107,7 @@ class CartService
             $cartItem = new CartItem();
             $cartItem->setProduct($this->variantRepository->find($id));
             $cartItem->setQuantity($quantity);
+            $cartItem->setIsPaid(false);
             $cartEntity->addCartItems($cartItem);
             $this->entityManager->persist($cartItem);
         }
@@ -147,7 +149,10 @@ class CartService
     }
     
     public function initCart(Cart $cartEntity) {
-        $cartEntity = $this->clearCartItems($cartEntity);
+        // foreach($cartEntity->getCartItems() as $cartItem) {
+        //     $cartItem->setIsPaid(true);
+        // }
+        $this->clearCartItems($cartEntity);
         $cartEntity->setTotalToPay(0);
         $cartEntity->setTotalTax(0);
         $cartEntity->setIsValidated(false);
@@ -158,8 +163,13 @@ class CartService
     private function clearCartItems(Cart $cartEntity) : ?Cart
     {
         foreach($cartEntity->getCartItems() as $cartItem) {
-            $cartEntity->removeCartItem($cartItem);
-            $this->entityManager->remove($cartItem);
+            if (!$cartItem->getOrder()) {
+                $cartEntity->removeCartItem($cartItem);
+                $this->entityManager->remove($cartItem);
+            } 
+            else {
+                $cartItem->setIsPaid(true);
+            }
         }
         $this->entityManager->flush();
         return $cartEntity;
